report 분리
