+++ eunphyo/week2/src/main/java/com/example/week2/report/dto/ReportRequest.java
@@ -1,4 +1,4 @@
-package com.example.week2.comment.dto;
+package com.example.week2.report.dto;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
@@ -7,7 +7,7 @@ import lombok.NoArgsConstructor;
 
 @Getter
 @NoArgsConstructor
-public class CommentReportRequest {
+public class ReportRequest {
 
     @NotNull(message = "신고자 ID는 필수입니다.")
     private Long userId;
