+++ eunphyo/week2/src/main/java/com/example/week2/report/dto/ReportResponse.java
@@ -1,0 +1,26 @@
+package com.example.week2.report.dto;
+
+import lombok.Builder;
+
+import java.time.LocalDateTime;
+
+public class ReportResponse {
+
+    @Builder
+    public record CommentReportResponse(
+            Long reportId,
+            Long commentId,
+            String reason,
+            String status,
+            LocalDateTime createdAt
+    ) {}
+
+    @Builder
+    public record CommentReportResolve(
+            Long reportId,
+            Long commentId,
+            String reason,
+            String status,
+            LocalDateTime createdAt
+    ) {}
+}
